Enhance security and billing with admin setup and BigDecimal migration

+++ src/main/java/io/github/ramiro/escapesj/servicio/FacturacionResult.java
@@ -1,12 +1,13 @@
 package io.github.ramiro.escapesj.servicio;
 
 import io.github.ramiro.escapesj.persistencia.BoletaRepository.BoletaItem;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FacturacionResult(
         int boletaId,
         int numero,
-        double subtotal,
-        double totalFinal,
+        BigDecimal subtotal,
+        BigDecimal totalFinal,
         List<BoletaItem> items
 ) {}
